added print func for aiding debugging xpath queries
added print function for aiding with debugging of xpath query
expressions. i.e. `print(//*)` it will log to the console the context
node, and value - in this example, it is a nodeset containing all nodes
in the document, so it will display the xpath of each node in the
console